Scraper/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
python-dotenv
beautifulsoup4
protobuf
sentencepiece